requirements.txt
flask==3.0.0
python-dotenv==1.0.0
langchain-community==0.2.16
langchain-ydb==0.0.4
yandexcloud==0.291.0
langchain==0.2.17
